Extract grouped window evaluation
